Built-RAG/requirements.txt
getpass
os
python-dotenv
langchain-core
ipython
langchain-google-genai
google-cloud-aiplatform
